Fixing bug which enables using floats for reconstruction

+++ Code/MBIRLib/Reconstruction/ReconstructionStructures.h
@@ -40,7 +40,7 @@
 #include <string>
 #include <vector>
 
-typedef double Real_t;
+typedef float Real_t;
 
 #include "MBIRLib/MBIRLib.h"
 #include "MBIRLib/Common/TomoArray.hpp"
